Spring Security JWT

+++ spring/Spring-11-Rest-Jackson/src/main/java/com/cybertek/controller/WebFluxController.java
@@ -101,7 +101,12 @@ public class WebFluxController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public Mono<Void> deleteWebClient(@PathVariable("id") Long id){
+    public Mono<Void> deleteWebClient(@PathVariable("id") Long id) throws Exception {
+        Integer countGenres = genreRepository.countGenresNativeQuery(id);
+        if(countGenres>0){
+            throw new Exception("Genre can't be deleted, is linked by a movie");
+        }
+
         return webClient.delete()
                 .uri("/delete-genre/{id}", id)
                 .retrieve()
